Fix: Handle low intel iGPU transcendental precision.
Intel iGPUs have terrible precision for sin/cos outside
of input range [-pi, pi) - at extremes, this results in
sin&cos returning one of 4 values: -1, -0.3, 0.3, 1.

See: https://stackoverflow.com/questions/21464664

The workaround is to mod the input argument so that it
is nearer the [-pi, pi) range (within an order of
magnitude should be fine).