[PlayOn] Add alignment force. Track occupied entitygrid cells.
It seems that alignment and separation are the two key steering behaviors
for large flocks. Previously entities would crawl into the ball via
walking a shell on the outside. Now, because of alignment units in the
center of the ball will shift until in essence the center of mass is at
the desired goal position.

Separation is probably too weak right now. Raw seeking behaviors are
aggregated into the alignment force. However, it's still necessary to use
them somehow (where the desired seek is in the opposite direction of the
alignment force, or where the alignment force is zeroed out but units are
still stuck away from the goal ball).

There are various tuning parameters that are certainly nonoptimal. The
biggest one is probablhy seekAlignWeight, which currently expects a
hardcoded ball radius of 200 units approx. It'd be better to count which
units are at goal and from there adjust that constant, or something
similar. The mix of seek forces (continuous vs discrete) isn't finalized
either.

This code is only dependant on position, goal, and statistics - no need
for prior frames' data. Doesn't matter if goal is respecified every frame
or not.

I still haven't figured out how fog of war should interact with flocking.
In essence, we need the goal to perform the flocking simulation. However,
exposing the goal to client is leaking information, since inn an ideal
world all the client knows is the position of the units and not where
they're heading.