[PlayOn] Prep work for path-to-hole/structure.
Moved PolyNodeCrossoverPointManager to own file.

Holes should likely be pathed to via crossover points (which should be
renamed into, say, landmark points). Crossover edges are still going to be
a thing, so some landmark points will be used to jump portals. I'll have
to brainstorm a better name.

A more generic approach is to register structures as both holes and
landmark regions, which can be pathed to. This is important because, for
example, pathing to water to go fishing might not care about eroding land
(it's okay to stand on the boundary? Maybe there's a better example) and
possibly might not require actually hole-punching (i.e. you can also walk
into the water).