[PlayOn] Implement ConvertTriangulationToSimplePolygon
This is part of implementing Section 3.1 of Zarei & Ghodsi (2008).

> Zarei, A., & Ghodsi, M. (2008). Query point visibility computation in polygons with holes. Computational Geometry, 39(2), 78–90. doi:10.1016/j.comgeo.2007.02.005

The algorithm works by building an MST of the triangulation where:

Nodes = Triangles
Edges = Shared Triangle Edges

Then, the MST is walked CCW to emit points along the simple polygon's
boundary. Future improvements might refine, say, the size of the results
buffer (can it be an array? It seems the output size is easily
precomputed?) or the MST heuristic used (currently just DFS).

Here're some output visualizations: