[terragami] correction to O(N) vispoly paper + some error tolerance
The paper's RETARD case A doesn't account for a new point being BEHIND
an existing stack line segment, -- the paper's code would cut this
segment. Instead, I've added a condition that ensures we only cut
segments if our next iteration point is in front of that segment.

You can repro this bug before this commit by running vispoly computation
against vert 8 of squiddy.

The paper also tests case a / case b in proc RETARD after scanning the
stack. This is more annoying for implementation, because you then have
to duplicate your test inside scanning stack to determine which branch
to execute. Instead, I've simply stored a var "are we doing case a or b"