[terragami] Track simple vispoly windows, reflex verts, segments, t.
Added lots of comments to explain what's going on.

For computing a visibility decomposition, we need to know the windows
in our vispoly & the segments they're associated with (e.g. so for
vertical lines we can start/split cells with the window segment).

One could prior have computed that through collinearity checks, though
this should be significantly faster (computed essentially for free) and
easier (don't need to associate windows with segments during line
sweep).

Note: The window rendering reveals a bug in our vispoly compute (well,
is it a bug or just a degenerate output?): In collinear cases, a window
endpoint vertex can be inserted right before an identical vertex.

I suspect the bug is in ADVANCE, and the fix is simply to check if the
prior point on the stack is loosely equivalent to the one we're about to
insert (and if so, to update it to the current point & have a stack
entry with both VertexIndex & WindowEndpointData)