[terragami] collinearity robustness for window stack entries
There are two cases I'm handling

1. We complete a SCAN by entering a point that is collinear to the
window. In this case, we don't want to add BOTH the window endpoint
(which SCAN will do) and the next vertex which ADVANCE would do, as
they should be merged. We fix this in ADVANCE - if the prior stack
entry is a window node & identical, merge the node-to-add into it.

2. We complete a case A RETARD by cutting segment s[j]-s[j+1]
with a new point, where the new point is collinear with v0 and s[j].
In this case, the window endpoint is at s[j], so instead of inserting
a new point, update the existing s[j] to include our window data.